COMPMID-3828: fix unsigned overflow in quantized reduce mean
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/4143
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>

+++ src/core/NEON/kernels/NEReductionOperationKernel.cpp
@@ -1459,7 +1459,7 @@ struct RedOpYZW_quantized
                     case ReductionOperation::MEAN_SUM:
                     {
                         int32_t res = static_cast<int32_t>(res_value);
-                        res /= in_info.dimension(axis);
+                        res /= static_cast<int32_t>(in_info.dimension(axis));
                         *reinterpret_cast<T *>(output.ptr() + x) = utils::cast::saturate_cast<T>(res);
                         break;
                     }
